refactor: move tr_num_parse_range() to remote.cc
It's the only code that uses it, so don't make everyone else pay for it.
